Time: 10 ms (41.58%), Space: 61.3 MB (84.95%) - LeetHub

+++ 2439-minimize-maximum-of-array/2439-minimize-maximum-of-array.java
@@ -4,7 +4,7 @@ class Solution {
         double total = nums[0];
         for (int i = 1; i < nums.length; i++) {
             total += nums[i];
-            res = (int) Math.max( (double) res, Math.ceil(total / (i + 1) ) );
+            res = (int) Math.max( res, Math.ceil(total / (i + 1) ) );
         }
         return res;
     }
